refactor: CustomQueryExecutionListener 위치 변경
